Have the ACD restart itself on configuration changes.
This is a non-optimal solution.  Currently the ACD doesn't know exactly
what changed, and so it restarts itself even if it may not have needed to.
Later this should be revisited to track changes and only require a restart
if the changed config cannot be handled "on the fly".

In addition, the need for a restart should be conveyed to sipXconfig
(via the return values of the XML-RPC, or a particular field to "get"), so
that it can get operator feedback if it is okay to restart the running ACD.
When that feedback is obtained, it can send a RESTART command.

Fixes XCF-1729, kinda sorta.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@10241 ab1d8caa-1f67-47f1-9e81-24633a41865c