sipXivr: Support for "None" MOH Audio Source option
Adds a new value to "n" to the MOH id.
   n        "None"

Causes the MOH server to accept the call, then hang up.  This should
end up sending no packets of MOH to the far end.

Another possiblity for "n" is to play silence.  But that causes
RTP packets to flow, it it isn't quite the same as "no music", it's
just really really quiet!  And John Cage's estate might have cause
to sue for copyright violation if the MOH session lasts for 4'33".

Fixes part of XX-7165 -  Add "None" MOH Audio Source option

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@17546 ab1d8caa-1f67-47f1-9e81-24633a41865c